fix: apply Suspense wrapper for SearchPage to resolve build error
